rename old Leo client package name to legacy_leonardo_client
