+++ api/src/main/java/org/pmiops/workbench/api/AppsController.java
@@ -5,8 +5,8 @@ import org.pmiops.workbench.config.WorkbenchConfig;
 import org.pmiops.workbench.db.model.DbUser;
 import org.pmiops.workbench.db.model.DbWorkspace;
 import org.pmiops.workbench.interactiveanalysis.InteractiveAnalysisService;
-import org.pmiops.workbench.leonardo.LeonardoApiClient;
-import org.pmiops.workbench.leonardo.LeonardoApiHelper;
+import org.pmiops.workbench.legacy_leonardo_client.LeonardoApiClient;
+import org.pmiops.workbench.legacy_leonardo_client.LeonardoApiHelper;
 import org.pmiops.workbench.model.AppLocalizeRequest;
 import org.pmiops.workbench.model.AppLocalizeResponse;
 import org.pmiops.workbench.model.CreateAppRequest;
