+++ api/src/main/java/org/pmiops/workbench/config/RetryConfig.java
@@ -3,7 +3,7 @@ package org.pmiops.workbench.config;
 import org.pmiops.workbench.calhoun.CalhounRetryHandler;
 import org.pmiops.workbench.firecloud.FirecloudRetryHandler;
 import org.pmiops.workbench.google.GoogleRetryHandler;
-import org.pmiops.workbench.leonardo.LeonardoRetryHandler;
+import org.pmiops.workbench.legacy_leonardo_client.LeonardoRetryHandler;
 import org.pmiops.workbench.notebooks.NotebooksRetryHandler;
 import org.pmiops.workbench.rawls.RawlsRetryHandler;
 import org.pmiops.workbench.shibboleth.ShibbolethRetryHandler;
