+++ api/src/main/java/org/pmiops/workbench/disks/DiskService.java
@@ -3,9 +3,9 @@ package org.pmiops.workbench.disks;
 import java.util.List;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
-import org.pmiops.workbench.leonardo.LeonardoApiClient;
-import org.pmiops.workbench.leonardo.PersistentDiskUtils;
-import org.pmiops.workbench.leonardo.model.LeonardoListPersistentDiskResponse;
+import org.pmiops.workbench.legacy_leonardo_client.LeonardoApiClient;
+import org.pmiops.workbench.legacy_leonardo_client.PersistentDiskUtils;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoListPersistentDiskResponse;
 import org.pmiops.workbench.model.Disk;
 import org.pmiops.workbench.utils.mappers.LeonardoMapper;
 import org.pmiops.workbench.workspaces.WorkspaceService;
