+++ api/src/main/java/org/pmiops/workbench/legacy_leonardo_client/LeonardoCustomEnvVarUtils.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.leonardo;
+package org.pmiops.workbench.legacy_leonardo_client;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.CharMatcher;
