+++ api/src/main/java/org/pmiops/workbench/legacy_leonardo_client/LeonardoLabelHelper.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.leonardo;
+package org.pmiops.workbench.legacy_leonardo_client;
 
 import jakarta.annotation.Nullable;
 import java.util.HashMap;
