+++ api/src/main/java/org/pmiops/workbench/legacy_leonardo_client/LeonardoRetryHandler.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.leonardo;
+package org.pmiops.workbench.legacy_leonardo_client;
 
 import jakarta.inject.Provider;
 import jakarta.servlet.http.HttpServletResponse;
