+++ api/src/main/java/org/pmiops/workbench/utils/mappers/LeonardoMapper.java
@@ -15,30 +15,30 @@ import org.mapstruct.MappingTarget;
 import org.mapstruct.Named;
 import org.mapstruct.ValueMapping;
 import org.pmiops.workbench.exceptions.ServerErrorException;
-import org.pmiops.workbench.leonardo.LeonardoLabelHelper;
-import org.pmiops.workbench.leonardo.model.LeonardoAllowedChartName;
-import org.pmiops.workbench.leonardo.model.LeonardoAppType;
-import org.pmiops.workbench.leonardo.model.LeonardoCloudContext;
-import org.pmiops.workbench.leonardo.model.LeonardoCloudProvider;
-import org.pmiops.workbench.leonardo.model.LeonardoClusterError;
-import org.pmiops.workbench.leonardo.model.LeonardoDiskConfig;
-import org.pmiops.workbench.leonardo.model.LeonardoDiskStatus;
-import org.pmiops.workbench.leonardo.model.LeonardoGceConfig;
-import org.pmiops.workbench.leonardo.model.LeonardoGceWithPdConfig;
-import org.pmiops.workbench.leonardo.model.LeonardoGetAppResponse;
-import org.pmiops.workbench.leonardo.model.LeonardoGetRuntimeResponse;
-import org.pmiops.workbench.leonardo.model.LeonardoKubernetesRuntimeConfig;
-import org.pmiops.workbench.leonardo.model.LeonardoListAppResponse;
-import org.pmiops.workbench.leonardo.model.LeonardoListPersistentDiskResponse;
-import org.pmiops.workbench.leonardo.model.LeonardoListRuntimeResponse;
-import org.pmiops.workbench.leonardo.model.LeonardoMachineConfig;
-import org.pmiops.workbench.leonardo.model.LeonardoPersistentDiskRequest;
-import org.pmiops.workbench.leonardo.model.LeonardoRuntimeConfig;
-import org.pmiops.workbench.leonardo.model.LeonardoRuntimeConfig.CloudServiceEnum;
-import org.pmiops.workbench.leonardo.model.LeonardoRuntimeImage;
-import org.pmiops.workbench.leonardo.model.LeonardoRuntimeStatus;
-import org.pmiops.workbench.leonardo.model.LeonardoUpdateDataprocConfig;
-import org.pmiops.workbench.leonardo.model.LeonardoUpdateGceConfig;
+import org.pmiops.workbench.legacy_leonardo_client.LeonardoLabelHelper;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoAllowedChartName;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoAppType;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoCloudContext;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoCloudProvider;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoClusterError;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoDiskConfig;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoDiskStatus;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoGceConfig;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoGceWithPdConfig;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoGetAppResponse;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoGetRuntimeResponse;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoKubernetesRuntimeConfig;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoListAppResponse;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoListPersistentDiskResponse;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoListRuntimeResponse;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoMachineConfig;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoPersistentDiskRequest;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoRuntimeConfig;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoRuntimeConfig.CloudServiceEnum;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoRuntimeImage;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoRuntimeStatus;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoUpdateDataprocConfig;
+import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoUpdateGceConfig;
 import org.pmiops.workbench.model.AppType;
 import org.pmiops.workbench.model.DataprocConfig;
 import org.pmiops.workbench.model.Disk;
